CodeReview: - Renamed project -> PluginBlockDetection
