+++ src/main/java/fr/nono74210/pluginblockdetection/commands/CommandReload.java
@@ -1,18 +1,18 @@
-package fr.nono74210.plugindetection.commands;
+package fr.nono74210.pluginblockdetection.commands;
 
-import fr.nono74210.plugindetection.MessageUtils;
+import fr.nono74210.pluginblockdetection.MessageUtils;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import fr.nono74210.plugindetection.PluginDetection;
+import fr.nono74210.pluginblockdetection.PluginBlockDetection;
 import org.jetbrains.annotations.NotNull;
 
 public class CommandReload implements CommandExecutor {
 
-    private final PluginDetection plugin;
+    private final PluginBlockDetection plugin;
 
     public CommandReload() {
-        plugin = PluginDetection.getInstance();
+        plugin = PluginBlockDetection.getInstance();
     }
 
     @Override
@@ -22,7 +22,7 @@ public class CommandReload implements CommandExecutor {
         }
 
         if (args.length != 1 || !args[0].equalsIgnoreCase("reload")) {
-            sender.sendMessage(MessageUtils.colorize(plugin.getConfig().getString("Messages.BadArguments", "&bPlugin&3Detection &8→ &cThere's an error in the command.")));
+            sender.sendMessage(MessageUtils.colorize(plugin.getConfig().getString("Messages.BadArguments", "&bPlugin&3Block&bDetection &8→ &cThere's an error in the command.")));
             return false;
         }
 
@@ -33,7 +33,7 @@ public class CommandReload implements CommandExecutor {
 
         plugin.reloadConfig();
 
-        sender.sendMessage(MessageUtils.colorize(plugin.getConfig().getString("Messages.ReloadComplete", "&bPlugin&3Detection &8→ &aPlugin reloaded without error.")));
+        sender.sendMessage(MessageUtils.colorize(plugin.getConfig().getString("Messages.ReloadComplete", "&bPlugin&3Block&bDetection &8→ &aPlugin reloaded without error.")));
         return true;
     }
